Consulted with lab instructor on workaround, added comments as per our discussion
I was having problems with concurrency, apparently "depends_on" in docker-compose did not always function as should. Added the following time out function as workaround to wait for postgres to start before client connects. Consulted this with lab instructor, he said this was OK and he said to keep the time out function here as is